Create 2. Biomass Analysis
Data was filtered to include only Forbs, Legumes and Graminoids, and "All", as there is inconsistency over the duration of the experiment in the measurement of the Woody, Dead and Bryophyte categories. The choice was also made that for the End harvest of Year 8, the biomass would not be sorted before drying, and so these weights are labeled with the fourth category "All".

When analysing data across the length of the treatment, Year 1 was removed. This is because the experimental treatments were only implemented in July of 2016, and so biomass collected for the Mid harvest, and potentially the End harvest of that year will be unlikely to be affected by the treatments. Year 1 was analysed as a year total and then each harvest separately, and indeed the total year biomass and the Mid harvest is not seen to have statistically significant effects of the treatments.