Create 3. Coefficient of Variance Analysis
As mentioned in document "2. Biomass Analysis", the data is filtered to Year 2 (2017) to Year 8 (2023), due to the experimental treatment being implemented July 2016. The data is also filtered for Forbs, Graminoids and Legumes (and "All), as the data collection of Bryophyte, Woody and Dead was not consistent


Aims: Identify any effects of precipitation manipulation on the Coefficient of Variance (stability metric).

Results: Drought is shown to have a statistically significant effect on the Coefficient of Variance of each plot, increasing CV and reducing stability. Additionally, plots show that the Drought treatment has the lowest temporal Mean and lowest temporal Standard Deviation.